requirements.txt
Django
django-extensions
django-admin-rangefilter
django-celery-beat
django-import-export
djangorestframework
djangorestframework-camel-case
djangorestframework-simplejwt
gunicorn
h11
hiredis
httplib2
httptools
msgpack
oauthlib
odfpy
openpyxl
protobuf
psycopg2-binary
pyasn1
pyasn1-modules
pycparser
PyJWT
pyOpenSSL
pyparsing
python-crontab
python-dateutil
python-dotenv
pytz
PyYAML
requests
requests-oauthlib
rsa
zope.interface
django-cors-headers
gevent
whitenoise

openai-whisper
setuptools-rust
speechbrain
pyannote.audio
whisper